backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
scikit-fuzzy>=0.4.2
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
networkx>=3.0
packaging
setuptools
psutil==5.9.8
google-genai==1.9.0
groq==1.1.0
python-dotenv==1.0.1
pydantic>=2.6.4
decorator
six
